Organizing single and multi RB walls
There are two folders depending on what you need.

Single RB is more reliable for level building and more optimized because it has less collision calculations.

Multi RB is cooler for the player, and perhaps designer wants to add a few of these around the level. These should be kept to a minimum and only for low-elevation walls. Overuse or attempts to place taller multiRB walls would slow down the game and the walls themselves would just crumble to bits.